move to shared: chat/{highlight,markdown}
