+++ client/cody-shared/src/chat/highlight.ts
