+++ client/cody-shared/src/chat/markdown.ts
