+++ client/cody/src/chat/recipes/generate-docstring.ts
@@ -1,3 +1,4 @@
+import { renderMarkdown } from '@sourcegraph/cody-shared/src/chat/markdown'
 import { Interaction } from '@sourcegraph/cody-shared/src/chat/transcript/interaction'
 import { IntentDetector } from '@sourcegraph/cody-shared/src/intent-detector'
 import { MAX_RECIPE_INPUT_TOKENS, MAX_RECIPE_SURROUNDING_TOKENS } from '@sourcegraph/cody-shared/src/prompt/constants'
@@ -6,7 +7,6 @@ import { getShortTimestamp } from '@sourcegraph/cody-shared/src/timestamp'
 
 import { CodebaseContext } from '../../codebase-context'
 import { Editor } from '../../editor'
-import { renderMarkdown } from '../markdown'
 
 import {
     MARKDOWN_FORMAT_PROMPT,
